Fixed table render issues

+++ app.py
@@ -55,12 +55,13 @@ def update(sno):
     return render_template("update.html", todo=todo)
 
 
-@app.route('/delete/<int:sno>')
+@app.route('/delete/<int:sno>', methods=["GET", "POST"])
 def delete(sno):
-    todo = Todo.query.filter_by(sno=sno).first()
-    db.session.delete(todo)
-    db.session.commit()
-    return redirect("/")
+    if request.method == "POST":
+        todo = Todo.query.filter_by(sno=sno).first()
+        db.session.delete(todo)
+        db.session.commit()
+        return redirect("/")
 
 
 if __name__ == "__main__":
